tabularis-server/pyproject.toml
[project]
name = "tabularis-server"
version = "0.1.0"
description = "FastAPI backend for Tabular — PDF to Excel conversion"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.30",
    "psycopg2-binary>=2.9",
    "python-jose[cryptography]>=3.3",
    "python-multipart>=0.0.9",
    "pdfplumber>=0.11",
    "openpyxl>=3.1",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
